+++ deno/lib/__tests__/array.test.ts
@@ -2,6 +2,7 @@
 import { expect } from "https://deno.land/x/expect@v0.2.6/mod.ts";
 const test = Deno.test;
 
+import { util } from "../helpers/util.ts";
 import * as z from "../index.ts";
 
 const minTwo = z.string().array().min(2);
@@ -9,6 +10,13 @@ const maxTwo = z.string().array().max(2);
 const justTwo = z.string().array().length(2);
 const intNum = z.string().array().nonempty();
 const nonEmptyMax = z.string().array().nonempty().max(2);
+
+type t1 = z.infer<typeof nonEmptyMax>;
+const f1: util.AssertEqual<[string, ...string[]], t1> = true;
+f1;
+type t2 = z.infer<typeof minTwo>;
+const f2: util.AssertEqual<string[], t2> = true;
+f2;
 
 test("passing validations", () => {
   minTwo.parse(["a", "a"]);
@@ -42,4 +50,18 @@ test("parse empty array in nonempty", () => {
 test("get element", () => {
   justTwo.element.parse("asdf");
   expect(() => justTwo.element.parse(12)).toThrow();
+});
+
+test("continue parsing despite array size error", () => {
+  const schema = z.object({
+    people: z.string().array().min(2),
+  });
+
+  const result = schema.safeParse({
+    people: [123],
+  });
+  expect(result.success).toEqual(false);
+  if (!result.success) {
+    expect(result.error.issues.length).toEqual(2);
+  }
 });