+++ deno/lib/__tests__/number.test.ts
@@ -4,21 +4,25 @@ const test = Deno.test;
 
 import * as z from "../index.ts";
 
-const minFive = z.number().min(5, "min5");
-const maxFive = z.number().max(5, "max5");
+const gtFive = z.number().gt(5);
+const gteFive = z.number().gte(5);
+const ltFive = z.number().lt(5);
+const lteFive = z.number().lte(5);
 const intNum = z.number().int();
 
 test("passing validations", () => {
-  minFive.parse(5);
-  minFive.parse(6);
-  maxFive.parse(5);
-  maxFive.parse(4);
+  gtFive.parse(6);
+  gteFive.parse(5);
+  ltFive.parse(4);
+  lteFive.parse(5);
   intNum.parse(4);
 });
 
 test("failing validations", () => {
-  expect(() => minFive.parse(4)).toThrow();
-  expect(() => maxFive.parse(6)).toThrow();
+  expect(() => ltFive.parse(5)).toThrow();
+  expect(() => lteFive.parse(6)).toThrow();
+  expect(() => gtFive.parse(5)).toThrow();
+  expect(() => gteFive.parse(4)).toThrow();
   expect(() => intNum.parse(3.14)).toThrow();
 });