+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+### 3.7
+
+- Eliminate `ZodNonEmptyArray`, add `Cardinality` to `ZodArray`
+- Add optional error message to `ZodArray.nonempty`
+- Add `.gt/.gte/.lt/.lte` to `ZodNumber`
+
 ### 3.6
 
 - Add IE11 support